added permissions checking and custom 403.html

+++ mysite/company/views.py
@@ -1,3 +1,5 @@
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.mixins import LoginRequiredMixin
 from django.core.exceptions import PermissionDenied
 from django.shortcuts import render, redirect
 # Create your views here.
@@ -14,11 +16,11 @@ from profile import models
 from profile.models import Manager, Profile
 
 
+@login_required
 def edit_company(request, company_id=None):
     if company_id:
         company_instance = Company.objects.get(pk=company_id)
-        if company_instance.manager != request.user.profile.manager:
-            raise PermissionDenied
+        check_manager_permission(company_instance.manager, request.user)
     else:
         company_instance = Company()
     company_form = CompanyForm(instance=company_instance)
@@ -50,11 +52,15 @@ def view_company(request, company_id):
                    'title': 'Company Details'})
 
 
+@login_required
 def remove_company(request, company_id):
+    company_instance = Company.objects.get(pk=company_id)
+    check_manager_permission(company_instance, request.user)
     Company.objects.get(pk=company_id).delete()
     return redirect('company-list')
 
 
+@login_required
 def view_my_companies(request):
     if Manager.objects.filter(profile=request.user.profile).exists():
         company_manager = request.user.profile.manager
@@ -71,8 +77,10 @@ def staff_has_service(staff_instance, service_id):
 
 
 # TODO[IM]: add errors to form
+@login_required
 def add_staff(request, company_id):
-    company = Company(pk=company_id)
+    company = Company.objects.get(pk=company_id)
+    check_manager_permission(company.manager, request.user)
     entered_username = None
     error_message = None
     if request.method == 'POST':
@@ -100,8 +108,10 @@ def add_staff(request, company_id):
                    'title': 'Add Staff'})
 
 
+@login_required
 def edit_staff(request, staff_id):
     staff_member = StaffMember.objects.get(pk=staff_id)
+    check_manager_permission(staff_member.company.manager, request.user)
     if request.method == 'POST':
         new_staff_service_list = request.POST.getlist('checked_services')
         staff_member.services.clear()
@@ -123,8 +133,10 @@ def view_staff(request, staff_id):
                   {'staff_member': StaffMember.objects.get(pk=staff_id), 'title': 'Staff Details'})
 
 
+@login_required
 def remove_staff(request, staff_id):
     staff_member = StaffMember.objects.get(pk=staff_id)
+    check_manager_permission(staff_member.company.manager, request.user)
     company_id = staff_member.company_id
     staff_member.delete()
     return redirect(reverse('staff-list', args=(company_id,)))
@@ -145,8 +157,10 @@ class StaffList(ListView):
         return context
 
 
+@login_required
 def edit_service(request, company_id, service_id=None):
     company = Company(pk=company_id)
+    check_manager_permission(company.manager, request.user)
     if service_id:
         service_instance = Service.objects.get(pk=service_id)
     else:
@@ -165,8 +179,10 @@ def edit_service(request, company_id, service_id=None):
                   {'service_form': service_form, 'title': 'Update Service'})
 
 
+@login_required
 def remove_service(request, service_id):
     service = Service.objects.get(pk=service_id)
+    check_manager_permission(service.company.manager, request.user)
     company_id = service.company_id
     service.delete()
     return redirect(reverse('service-list', args=(company_id,)))
@@ -192,12 +208,13 @@ def view_service(request, service_id):
                   {'service': Service.objects.get(pk=service_id), 'title': 'Service Details'})
 
 
-class CompanyOrderList(ListView):
+class CompanyOrderList(LoginRequiredMixin, ListView):
     model = Order
     template_name = "company/company/company_orders.html"
     paginate_by = 9
 
     def get_queryset(self):
+        check_manager_permission(Company.objects.get(pk=self.kwargs['company_id']).manager, self.request.user)
         if self.kwargs['filter'] != 'all':
             return Order.objects.filter(service_order__company_id=self.kwargs['company_id'],
                                         order_state=self.kwargs['filter']).order_by('-date_created')
@@ -213,11 +230,12 @@ class CompanyOrderList(ListView):
         return context
 
 
+@login_required
 def update_company_orders(request, company_id):
     if request.method == 'POST':
         order_id = request.POST.get('order_id')
         order_obj = Order.objects.get(pk=order_id)
-
+        check_manager_permission(order_obj.service_order.company.manager, request.user)
         if 'but1' in request.POST:
             order_obj.order_state = Order.OrderState.approved
             order_obj.save()
@@ -248,6 +266,7 @@ class CompanyList(ListView):
         return context
 
 
+@login_required
 def add_comment(request, company_id):
     url = request.META.get('HTTP_REFERER')
     if request.method == 'POST':  # check post
@@ -264,3 +283,11 @@ def add_comment(request, company_id):
 
     comments_list = Comment.objects.all().filter(company=Company.objects.get(pk=company_id))
     return render(request, "company/company/review_message.html", {'comments_list': comments_list})
+
+
+def check_manager_permission(company_manager, user):
+    try:
+        if company_manager != user.profile.manager:
+            raise PermissionDenied
+    except Exception:
+        raise PermissionDenied
